+++ service/src/main/java/com/gamesky/card/service/key/CheckCodeKey.java
@@ -2,7 +2,6 @@ package com.gamesky.card.service.key;
 
 import com.gamesky.card.core.Cacheable;
 import com.gamesky.card.core.Constants;
-import com.gamesky.card.core.model.User;
 
 /**
  * Created on 6/11/15.
@@ -12,13 +11,13 @@ import com.gamesky.card.core.model.User;
 public class CheckCodeKey implements Cacheable {
 
     private final String phone;
-    private long expire = 60;   //秒
+    private int expire = 60;   //秒
 
     public CheckCodeKey(String phone) {
         this.phone = phone;
     }
 
-    public CheckCodeKey(String phone, long expire) {
+    public CheckCodeKey(String phone, int expire) {
         this(phone);
         this.expire = expire;
     }
@@ -29,7 +28,7 @@ public class CheckCodeKey implements Cacheable {
     }
 
     @Override
-    public long expire() {
+    public int expire() {
         return expire;
     }
 }
